# shifter/shifter_platform/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shifter Platform API",
    "version": "v1",
    "description": "Authenticated HTTP/JSON API for Shifter platform integrations."
  },
  "paths": {
...
    "/api/v1/mission-control/guacamole/bootstrap/{request_id}/": {
      "get": {
        "operationId": "api_v1_mission_control_guacamole_bootstrap_status_retrieve",
        "description": "Return current status for a queued Guacamole bootstrap request.",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuacamoleBootstrapStatus"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found."
          }
        },
        "x-required-scopes": [
          "mission_control:guacamole:read"
        ]
      }
    },
    "/api/v1/mission-control/guacamole/bootstrap/{request_id}/open/": {
      "get": {
        "operationId": "api_v1_mission_control_guacamole_bootstrap_open_retrieve",
        "description": "Render an opener page that polls until the bootstrap URL is ready.",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:guacamole:read"
        ]
      }
    },
    "/api/v1/mission-control/guacamole/rdp-url/": {
      "post": {
        "operationId": "api_v1_mission_control_guacamole_rdp_url",
        "description": "Queue an RDP bootstrap request for a range instance.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuacamoleInstance"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/GuacamoleInstance"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/GuacamoleInstance"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuacamoleBootstrapQueued"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:guacamole:read"
        ]
      }
    },
    "/api/v1/mission-control/guacamole/ssh-url/": {
      "post": {
        "operationId": "api_v1_mission_control_guacamole_ssh_url",
        "description": "Queue an SSH bootstrap request for a range instance.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuacamoleInstance"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/GuacamoleInstance"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/GuacamoleInstance"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuacamoleBootstrapQueued"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:guacamole:read"
        ]
      }
    },
...
    "/api/v1/mission-control/ngfw/{app_id}/ssh-url/": {
      "post": {
        "operationId": "api_v1_mission_control_ngfw_ssh_url",
        "description": "Queue an SSH bootstrap request for an NGFW instance.",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuacamoleBootstrapQueued"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found."
          }
        },
        "x-required-scopes": [
          "mission_control:guacamole:read"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Schema-only component for the canonical platform API error envelope.\n\nMirrors :func:`shared.api.errors.api_exception_handler` /\n:func:`shared.api.errors.api_error_response` output. It describes the wire\ncontract; it never runs at request time.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "Schema-only description of the ``shared.api.errors`` envelope body.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Safe, user-facing error message."
          },
          "details": {
            "description": "Optional structured field-level detail (present on validation errors)."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation id echoed from the X-Request-ID request header, when present."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "GuacamoleBootstrapQueued": {
        "type": "object",
        "description": "Response body for a queued Guacamole bootstrap (RDP/SSH/NGFW-SSH POST, HTTP 202).\n\n``url`` here is the compatibility opener route, not a signed session URL\n(that one-time URL is delivered only by the status poll, see\n``GuacamoleBootstrapStatusSerializer``).",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "status_url": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "status",
          "status_url",
          "url"
        ]
      },
      "GuacamoleBootstrapStatus": {
        "type": "object",
        "description": "Response body for ``GuacamoleBootstrapStatusView.get``.\n\n``url`` is the one-time signed Guacamole session URL, delivered exactly\nonce on a ``succeeded`` poll; never given a real example. ``duration_ms``\nand ``error`` are present only once the bootstrap has finished or failed.",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "duration_ms": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "status"
        ]
      },
      "GuacamoleInstance": {
        "type": "object",
        "description": "Validate range Guacamole URL bootstrap requests.",
        "properties": {
          "instance_uuid": {
            "type": "string"
          }
        },
        "required": [
          "instance_uuid"
        ]
      },
...
    },
    "securitySchemes": {
      "ApiTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "shf",
        "description": "Platform API token with scopes from shared.api_tokens.scopes."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  }
}